Extract statistics gathering into reusable utility
Reduces ~300 lines of duplicated code by creating a centralized
statistics utility function. This refactoring:

- Creates src/api/statistics-utils.ts with getRunStatsSnapshot()
- Defines DEFAULT_RECENT_STATS constant for type-safe defaults
- Replaces duplicated patterns in mobile and desktop layouts
- Improves maintainability and type consistency

No behavioral changes - pure internal refactoring.
